Develop

+++ app/Http/Controllers/ListingController.php
@@ -20,7 +20,8 @@ use Illuminate\Validation\Rule;
 
 class ListingController extends Controller
 {
-    public function index() {
+    public function index()
+    {
         return view('listings.index', [
             'listings' => Listing::latest()
                 ->filter(request(['tag', 'search']))
@@ -28,19 +29,22 @@ class ListingController extends Controller
         ]);
     }
 
-    public function show(Listing $listing) {
+    public function show(Listing $listing)
+    {
         return view('listings.show', [
             'listing' => $listing
         ]);
     }
 
     // Show create form
-    public function create() {
+    public function create()
+    {
         return view('listings.create');
     }
 
     // Store listing data
-    public function store(Request $request) {
+    public function store(Request $request)
+    {
         $form_fields = $request->validate([
             'title' => 'required',
             'company' => ['required', Rule::unique('listings', 'company')],
@@ -63,11 +67,18 @@ class ListingController extends Controller
         return redirect('/')->with('message', 'Job Listing Created Successfully!');
     }
 
-    public function edit(Listing $listing) {
+    public function edit(Listing $listing)
+    {
+        // Make sure logged in user is owner of listing
+        if ($listing->user_id != auth()->id()) {
+            return redirect('/')->with('message', 'You can not edit other users listings!');
+        }
+
         return view('listings.edit', ['listing' => $listing]);
     }
 
-    public function update(Request $request, Listing $listing) {
+    public function update(Request $request, Listing $listing)
+    {
         $form_fields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -87,9 +98,20 @@ class ListingController extends Controller
         return back()->with('message', 'Job Listing Updated Successfully!');
     }
 
-    public function destroy(Listing $listing) {
+    public function destroy(Listing $listing)
+    {
+        // Make sure logged in user is owner of listing
+        if ($listing->user_id != auth()->id()) {
+            return redirect('/')->with('message', 'You can not edit other users listings!');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Job Listing Deleted Successfully!');
+    }
+
+    // Manage Listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
